requirements.txt
Flask-xCaptcha==0.5.4
plotly==5.10.0
